requirements.txt
jax
jaxlib
flax
optax
hydra-core
wandb
matplotlib
numpy
scipy
absl-py
ml_collections
